Rename module for GitHub issue
The module has been renamed to indicate that the issue was created in
the crates.io repository. This might be relevant in the future, when
more smoke tests are added that might've been reported in other
repositories.
