+++ src/crates/crates_4891/cloudfront_space.rs
@@ -1,4 +1,4 @@
-//! Test CloudFront with an un-encoded URL
+//! Test CloudFront with a URL including a space character
 
 use async_trait::async_trait;
 use reqwest::StatusCode;
@@ -9,18 +9,18 @@ use super::config::Config;
 use super::request_url_and_expect_status;
 
 /// The name of the test
-const NAME: &str = "CloudFront unencoded";
+const NAME: &str = "CloudFront with space";
 
-/// Test CloudFront with an un-encoded URL
+/// Test CloudFront with a URL including a space character
 ///
-/// This test request a URL with an un-encoded `+` character from Cloudfront. The test expects the
-/// CDN to return an HTTP 200 OK response.
-pub struct CloudfrontUnencoded<'a> {
+/// This test request a URL with a space character from Cloudfront. The test expects the CDN to
+/// return an HTTP 403 Forbidden response.
+pub struct CloudfrontSpace<'a> {
     /// Configuration for this test
     config: &'a Config,
 }
 
-impl<'a> CloudfrontUnencoded<'a> {
+impl<'a> CloudfrontSpace<'a> {
     /// Create a new instance of the test
     pub fn new(config: &'a Config) -> Self {
         Self { config }
@@ -28,7 +28,7 @@ impl<'a> CloudfrontUnencoded<'a> {
 }
 
 #[async_trait]
-impl<'a> Test for CloudfrontUnencoded<'a> {
+impl<'a> Test for CloudfrontSpace<'a> {
     async fn run(&self) -> TestResult {
         let url = format!(
             "{}/crates/{}/{}-{}.crate",
@@ -36,35 +36,38 @@ impl<'a> Test for CloudfrontUnencoded<'a> {
             self.config.krate(),
             self.config.krate(),
             self.config.version()
-        );
+        )
+        .replace('+', " ");
 
-        request_url_and_expect_status(NAME, &url, StatusCode::OK).await
+        request_url_and_expect_status(NAME, &url, StatusCode::FORBIDDEN).await
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::crates::issue_4891::tests::setup;
+    use crate::crates::crates_4891::tests::setup;
     use crate::test_utils::*;
 
     use super::*;
 
     const KRATE: &str = "rust-cratesio-4891";
-    const VERSION: &str = "0.1.0+1";
+    const VERSION: &str = "0.1.0 1";
 
     #[tokio::test]
-    async fn succeeds_with_http_200_response() {
+    async fn succeeds_with_http_403_response() {
         let (mut server, config) = setup(KRATE, VERSION).await;
+
+        let encoded_version = VERSION.replace(' ', "%20");
 
         let mock = server
             .mock(
                 "GET",
-                format!("/crates/{KRATE}/{KRATE}-{VERSION}.crate").as_str(),
+                format!("/crates/{KRATE}/{KRATE}-{encoded_version}.crate").as_str(),
             )
-            .with_status(200)
+            .with_status(403)
             .create();
 
-        let result = CloudfrontUnencoded::new(&config).run().await;
+        let result = CloudfrontSpace::new(&config).run().await;
 
         // Assert that the mock was called
         mock.assert();
@@ -76,15 +79,17 @@ mod tests {
     async fn fails_with_other_http_responses() {
         let (mut server, config) = setup(KRATE, VERSION).await;
 
+        let encoded_version = VERSION.replace(' ', "%20");
+
         let mock = server
             .mock(
                 "GET",
-                format!("/crates/{KRATE}/{KRATE}-{VERSION}.crate").as_str(),
+                format!("/crates/{KRATE}/{KRATE}-{encoded_version}.crate").as_str(),
             )
-            .with_status(403)
+            .with_status(200)
             .create();
 
-        let result = CloudfrontUnencoded::new(&config).run().await;
+        let result = CloudfrontSpace::new(&config).run().await;
 
         // Assert that the mock was called
         mock.assert();
@@ -94,16 +99,16 @@ mod tests {
 
     #[test]
     fn trait_send() {
-        assert_send::<CloudfrontUnencoded>();
+        assert_send::<CloudfrontSpace>();
     }
 
     #[test]
     fn trait_sync() {
-        assert_sync::<CloudfrontUnencoded>();
+        assert_sync::<CloudfrontSpace>();
     }
 
     #[test]
     fn trait_unpin() {
-        assert_unpin::<CloudfrontUnencoded>();
+        assert_unpin::<CloudfrontSpace>();
     }
 }
